Refactor ServerPasswordModifyHandler to abstract core logic into a service class. Makes the handler more focused on its concerns. Also move all password hashing into a password hash service.
Move markMustChangePassword() and clearMustChangePassword() onto AuthenticatedTokenInterface.
