+++ src/FreeDSx/Ldap/Protocol/ServerProtocolHandler/ServerPasswordModifyHandler.php
@@ -15,10 +15,7 @@ namespace FreeDSx\Ldap\Protocol\ServerProtocolHandler;
 
 use FreeDSx\Asn1\Exception\EncoderException;
 use FreeDSx\Ldap\Control\Control;
-use FreeDSx\Ldap\Control\PwdPolicyError;
-use FreeDSx\Ldap\Entry\Change;
 use FreeDSx\Ldap\Entry\Dn;
-use FreeDSx\Ldap\Entry\Entry;
 use FreeDSx\Ldap\Exception\OperationException;
 use FreeDSx\Ldap\Operation\LdapResult;
 use FreeDSx\Ldap\Operation\Request\ExtendedRequest;
@@ -29,29 +26,17 @@ use FreeDSx\Ldap\Protocol\Factory\ResponseFactory;
 use FreeDSx\Ldap\Protocol\LdapMessageRequest;
 use FreeDSx\Ldap\Protocol\LdapMessageResponse;
 use FreeDSx\Ldap\Protocol\Queue\ServerQueue;
-use FreeDSx\Ldap\Server\AccessControl\AccessControlInterface;
-use FreeDSx\Ldap\Server\AccessControl\OperationType;
-use FreeDSx\Ldap\Server\Backend\Auth\NameResolver\BindNameResolverInterface;
-use FreeDSx\Ldap\Server\Backend\Auth\PasswordHasher;
-use FreeDSx\Ldap\Server\Backend\Auth\PasswordHashVerifier;
-use FreeDSx\Ldap\Server\Backend\LdapBackendInterface;
-use FreeDSx\Ldap\Server\Backend\Write\Command\UpdateCommand;
-use FreeDSx\Ldap\Server\Backend\Write\WriteContext;
-use FreeDSx\Ldap\Server\Backend\Write\WriteOperationDispatcher;
 use FreeDSx\Ldap\Server\Logging\EventContext;
 use FreeDSx\Ldap\Server\Logging\EventLogger;
 use FreeDSx\Ldap\Server\Logging\ServerEvent;
-use FreeDSx\Ldap\Server\PasswordPolicy\Attempt\PasswordModifyAttempt;
-use FreeDSx\Ldap\Server\PasswordPolicy\Guard\PasswordPolicyChangeGuard;
-use FreeDSx\Ldap\Server\PasswordPolicy\Decision\OperationalChanges;
-use FreeDSx\Ldap\Server\PasswordPolicy\Decision\PasswordPolicyOutcome;
+use FreeDSx\Ldap\Server\PasswordModify\PasswordModifyResult;
+use FreeDSx\Ldap\Server\PasswordModify\PasswordModifyService;
 use FreeDSx\Ldap\Server\PasswordPolicy\PasswordPolicyContext;
 use FreeDSx\Ldap\Server\Token\AuthenticatedTokenInterface;
-use FreeDSx\Ldap\Server\Token\BindToken;
 use FreeDSx\Ldap\Server\Token\TokenInterface;
 
 /**
- * Handles RFC 3062 Password Modify extended requests.
+ * Adapts RFC 3062 Password Modify requests to {@see PasswordModifyService}: decode, delegate, encode the response.
  *
  * @author Chad Sikorra <Chad.Sikorra@gmail.com>
  */
@@ -61,15 +46,9 @@ readonly class ServerPasswordModifyHandler implements ServerProtocolHandlerInter
 
     public function __construct(
         private ServerQueue $queue,
-        private LdapBackendInterface $backend,
-        private WriteOperationDispatcher $writeDispatcher,
-        private AccessControlInterface $accessControl,
-        private BindNameResolverInterface $identityResolver,
+        private PasswordModifyService $service,
         private EventLogger $eventLogger = new EventLogger(null),
-        private PasswordHashVerifier $hashVerifier = new PasswordHashVerifier(),
-        private PasswordHasher $hasher = new PasswordHasher(),
         private ResponseFactory $responseFactory = new ResponseFactory(),
-        private ?PasswordPolicyChangeGuard $changeGuard = null,
         private ?PasswordPolicyContext $passwordPolicyContext = null,
     ) {}
 
@@ -87,9 +66,14 @@ readonly class ServerPasswordModifyHandler implements ServerProtocolHandlerInter
 
         try {
             $this->assertNoCriticalUnsupportedControls($message->controls());
-            $targetDn = $this->process(
+            $result = $this->changePassword(
                 $message,
                 $token,
+            );
+            $targetDn = $result->targetDn;
+            $this->sendSuccess(
+                $message,
+                $result,
             );
         } catch (OperationException $e) {
             $control = $this->passwordPolicyControl();
@@ -122,12 +106,11 @@ readonly class ServerPasswordModifyHandler implements ServerProtocolHandlerInter
 
     /**
      * @throws OperationException
-     * @throws EncoderException
      */
-    private function process(
+    private function changePassword(
         LdapMessageRequest $message,
         TokenInterface $token,
-    ): Dn {
+    ): PasswordModifyResult {
         if (!$token instanceof AuthenticatedTokenInterface) {
             throw new OperationException(
                 'Authentication required.',
@@ -137,107 +120,27 @@ readonly class ServerPasswordModifyHandler implements ServerProtocolHandlerInter
 
         /** @var ExtendedRequest $raw */
         $raw = $message->getRequest();
-        $request = PasswordModifyRequest::fromAsn1($raw->toAsn1());
-        $entry = $this->resolveTargetEntry(
-            $request,
+
+        return $this->service->change(
+            PasswordModifyRequest::fromAsn1($raw->toAsn1()),
             $token,
+            $message->controls(),
         );
-        $targetDn = $entry->getDn();
+    }
 
-        $this->authorizeRequest(
-            $token,
-            $targetDn,
-        );
-        $this->verifyOldPassword(
-            $request,
-            $entry,
-        );
-
-        $newPassword = $request->getNewPassword();
-        $generated = null;
-
-        if ($newPassword === null) {
-            $generated = $this->hasher->generate();
-            $newPassword = $generated;
-        }
-
-        $isSelf = $this->isSelf(
-            $token,
-            $targetDn,
-        );
-        $this->assertSelfChangeWhenMustChange(
-            $token,
-            $isSelf,
-        );
-        $hashed = $this->hasher->hash($newPassword);
-        $deltas = $this->changeGuard?->enforce(new PasswordModifyAttempt(
-            target: $entry,
-            newPassword: $newPassword,
-            hashedNewPassword: $hashed,
-            oldPassword: $request->getOldPassword(),
-            isSelf: $isSelf,
-            passwordIsCleartext: true,
-        )) ?? OperationalChanges::none();
-
-        $this->applyPasswordChange(
-            $targetDn,
-            $hashed,
-            $deltas,
-            $token,
-            $message,
-        );
-
-        // A successful self-change satisfies any pwdReset requirement, so lift the session restriction immediately.
-        if ($isSelf && $token instanceof BindToken) {
-            $token->clearMustChangePassword();
-        }
-
+    private function sendSuccess(
+        LdapMessageRequest $message,
+        PasswordModifyResult $result,
+    ): void {
         $control = $this->passwordPolicyControl();
         $this->queue->sendMessage(new LdapMessageResponse(
             $message->getMessageId(),
             new PasswordModifyResponse(
                 new LdapResult(ResultCode::SUCCESS),
-                $generated,
+                $result->generatedPassword,
             ),
             ...($control === null ? [] : [$control]),
         ));
-
-        return $targetDn;
-    }
-
-    private function isSelf(
-        AuthenticatedTokenInterface $token,
-        Dn $targetDn,
-    ): bool {
-        return $token->getResolvedDn()->normalize()->toString() === $targetDn->normalize()->toString();
-    }
-
-    /**
-     * A pwdReset identity may only change its own password.
-     *
-     * An exop targeting another entry is refused (draft-behera-10 §8.1.2).
-     *
-     * @throws OperationException
-     */
-    private function assertSelfChangeWhenMustChange(
-        AuthenticatedTokenInterface $token,
-        bool $isSelf,
-    ): void {
-        if ($isSelf || !$token->mustChangePassword()) {
-            return;
-        }
-
-        $outcome = PasswordPolicyOutcome::deny(
-            PwdPolicyError::CHANGE_AFTER_RESET,
-            ResultCode::UNWILLING_TO_PERFORM,
-            'The password must be changed before any other operation is permitted.',
-        );
-        $this->passwordPolicyContext?->setOutcome($outcome);
-
-        throw new OperationException(
-            $outcome->diagnostic,
-            $outcome->ldapResultCode,
-        );
     }
 
     private function passwordPolicyControl(): ?Control
@@ -276,113 +179,6 @@ readonly class ServerPasswordModifyHandler implements ServerProtocolHandlerInter
             $context,
             subject: $token,
             message: $message,
-        );
-    }
-
-    /**
-     * @throws OperationException
-     */
-    private function resolveTargetEntry(
-        PasswordModifyRequest $request,
-        AuthenticatedTokenInterface $token,
-    ): Entry {
-        $userIdentity = $request->getUsername();
-
-        $entry = $userIdentity === null || $userIdentity === ''
-            ? $this->backend->get($token->getResolvedDn())
-            : $this->identityResolver->resolve(
-                $userIdentity,
-                $this->backend,
-            );
-
-        if ($entry === null) {
-            throw new OperationException(
-                'The target entry does not exist.',
-                ResultCode::NO_SUCH_OBJECT,
-            );
-        }
-
-        return $entry;
-    }
-
-    /**
-     * @throws OperationException
-     */
-    private function authorizeRequest(
-        AuthenticatedTokenInterface $token,
-        Dn $targetDn,
-    ): void {
-        $this->accessControl->authorizeOperation(
-            OperationType::PasswordModify,
-            $token,
-            $targetDn,
-        );
-        $this->accessControl->authorizeAttribute(
-            $token,
-            $targetDn,
-            'userPassword',
-        );
-    }
-
-    /**
-     * @throws OperationException
-     */
-    private function verifyOldPassword(
-        PasswordModifyRequest $request,
-        Entry $entry,
-    ): void {
-        $oldPassword = $request->getOldPassword();
-
-        if ($oldPassword === null) {
-            return;
-        }
-
-        foreach ($entry->get('userPassword')?->getValues() ?? [] as $stored) {
-            if ($this->hashVerifier->verify($oldPassword, $stored)) {
-                return;
-            }
-        }
-
-        throw new OperationException(
-            'Invalid credentials.',
-            ResultCode::INVALID_CREDENTIALS,
-        );
-    }
-
-    /**
-     * Persists the new password plus any policy bookkeeping deltas in one command. When deltas are present the
-     * command is system-flagged so the NO-USER-MODIFICATION operational attributes are accepted; ACL was already
-     * checked in authorizeRequest().
-     *
-     * @throws OperationException
-     */
-    private function applyPasswordChange(
-        Dn $targetDn,
-        string $hashed,
-        OperationalChanges $deltas,
-        TokenInterface $token,
-        LdapMessageRequest $message,
-    ): void {
-        $changes = [
-            Change::replace('userPassword', $hashed),
-            ...$deltas->changes,
-        ];
-        $context = $deltas->isEmpty()
-            ? new WriteContext(
-                $token,
-                $message->controls(),
-            )
-            : WriteContext::system(
-                $token,
-                $message->controls(),
-            );
-
-        $this->writeDispatcher->dispatch(
-            new UpdateCommand(
-                $targetDn,
-                $changes,
-            ),
-            $context,
         );
     }
 }
