+++ src/FreeDSx/Ldap/Server/Backend/Auth/PasswordPolicyAwareAuthenticator.php
@@ -22,7 +22,6 @@ use FreeDSx\Ldap\Server\PasswordPolicy\Guard\PasswordPolicyBindGuard;
 use FreeDSx\Ldap\Server\PasswordPolicy\PasswordPolicyResolver;
 use FreeDSx\Ldap\Server\PasswordPolicy\UserPasswordState;
 use FreeDSx\Ldap\Server\Token\AuthenticatedTokenInterface;
-use FreeDSx\Ldap\Server\Token\BindToken;
 use FreeDSx\Sasl\Mechanism\MechanismName;
 use SensitiveParameter;
 
@@ -86,7 +85,7 @@ final readonly class PasswordPolicyAwareAuthenticator implements PasswordAuthent
 
         $this->guard->recordSuccess($attempt);
 
-        if ($attempt->state->mustChange && $token instanceof BindToken) {
+        if ($attempt->state->mustChange) {
             $token->markMustChangePassword();
         }
