+++ src/FreeDSx/Ldap/Server/PasswordPolicy/Constraint/HistoryConstraint.php
@@ -15,7 +15,7 @@ namespace FreeDSx\Ldap\Server\PasswordPolicy\Constraint;
 
 use FreeDSx\Ldap\Control\PwdPolicyError;
 use FreeDSx\Ldap\Operation\ResultCode;
-use FreeDSx\Ldap\Server\Backend\Auth\PasswordHashVerifier;
+use FreeDSx\Ldap\Server\Backend\Auth\PasswordHashService;
 use FreeDSx\Ldap\Server\PasswordPolicy\Decision\PasswordPolicyOutcome;
 
 /**
@@ -23,10 +23,10 @@ use FreeDSx\Ldap\Server\PasswordPolicy\Decision\PasswordPolicyOutcome;
  */
 final readonly class HistoryConstraint implements PasswordChangeConstraint
 {
-    public function __construct(private PasswordHashVerifier $hashVerifier) {}
+    public function __construct(private PasswordHashService $hashService) {}
 
     /**
-     * Comparison goes through {@see PasswordHashVerifier}; entries stored under a no-longer-supported
+     * Comparison goes through {@see PasswordHashService}; entries stored under a no-longer-supported
      * scheme silently fail to match and effectively drop out of the history set (draft-behera-10 §5.3.7).
      */
     public function check(PasswordChangeAttempt $attempt): ?PasswordPolicyOutcome
@@ -37,7 +37,7 @@ final readonly class HistoryConstraint implements PasswordChangeConstraint
         }
 
         foreach ($attempt->state->history as $entry) {
-            if ($this->hashVerifier->verify($attempt->newPassword, $entry->data)) {
+            if ($this->hashService->verify($attempt->newPassword, $entry->data)) {
                 return PasswordPolicyOutcome::deny(
                     PwdPolicyError::PASSWORD_IN_HISTORY,
                     ResultCode::CONSTRAINT_VIOLATION,
